mcps/gmail/requirements.txt
mcp>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
starlette>=0.40.0
uvicorn>=0.30.0
